donut 3d
new algo

pcl registraiton exapmple

eigen added

added the open3d

added a filegroup

able to visualize something

added the fmt and eifgen to open3d build

chnages in open3d

addded python support

added the pip install numpy

added muliple

commneting out the compiling pip for lnot updating the pip

changed src to requirements in

adding matplotlib

updated python repo

matplot lib can visulize

added matplotlib

applesupport

docker file

docker file for workspace

docker file update

rules pcl for mac support

eigen examples

python example

python requirements update

grdient testing

changes

opencv mac support

opencv mac support

fixed the macos linux problem without cuda and open3d and glfw

cuda flag added

added glew

added eules vulkan

rules vulkan added

imgui works

metal works

changing the header

rules vulkan now added for automatic download

+++ src/opencv/line_detect/LineFitting.cpp
@@ -90,7 +90,7 @@ std::vector<cv::Vec2f> LineFitting::FitLine(const cv::Mat &input_image) {
     const int r = line_index - (n + 1) * (number_of_rhos_ + 2) - 1;
     const double rho = (r - (number_of_rhos_ - 1) * 0.5f) * rho_step_;
     const double angle = minimum_theta_ + n * theta_step_;
-    lines.emplace_back(cv::Vec2f{rho, angle});
+    lines.emplace_back(cv::Vec2f{static_cast<float>(rho), static_cast<float>(angle)});
   }
   return lines;
 }
